fix: Add packages needed for initramfs + root passwd
We turned off most of dracut's host detection, so it depends on these 
packages being available in its build process to successfully generate
a initramfs that switches to the real root.

Add a temporary root "changeme" password just so we can login for dev
purposes.In production we should use build-args or something more 
robust.

Signed-off-by: Sam Dasilva <samuelramos852@gmail.com>